Add a rake task to help generate the email
This script is designed to simplify my workflow. I'll write the blog
post in jekyll first, and then run `rake tl` to copy the HTML of the
content of the post to the clipboard. The HTML that jekyll generates is
pretty clean, and tinyletter's "source" view in the wysiwyg editor seems
to like it so this should make things easier.